tool is now usable from command line

+++ data-clustering/cluster.py
@@ -5,12 +5,11 @@ from numpy import array
 import numpy as np
 from scipy.cluster.vq import kmeans
 import os
+import sys
+import getopt
 
-input_csv = "/Users/jason/Documents/UCSC/CS198/POP_GDP_MERGE_ENERGY_PER.csv"
 meta_file = "meta.csv"
 output_csv = "out.csv"
-final_outfile = "final.csv"
-clusters = 3
 cmin = 1
 cmax = 10
 
@@ -32,7 +31,7 @@ def cluster(df, means, csv_min, csv_max):
         data.append(np.sort(kmeans(b.astype(np.float), means)[0]))
     return data
 
-def reshape(output_csv, clusters):
+def reshape(output_csv, clusters, final_outfile, input_csv):
     final_output = csv.writer(file(final_outfile, 'w+'))
     df = pd.read_csv(file(output_csv, 'r'))
     to_zip = []
@@ -48,7 +47,19 @@ def reshape(output_csv, clusters):
         outstring[:0] = [row]
         final_output.writerow(outstring)
 
-def main():
+def main(argv):
+    try:
+        optlist, args = getopt.getopt(argv, 'i:o:c:')
+    except getopt.GetoptError as err:
+        print str(err)
+        usage()
+        sys.exit(2)
+    arguments = map(list, zip(*optlist))[1]
+    input_csv = arguments[0]
+    final_outfile = arguments[1]
+    clusters = int(arguments[2])
+
+    cleanCSV(input_csv)
     df = pd.read_csv(file(meta_file, 'r'))
     out = file(output_csv, 'w+')
     output = csv.writer(out)
@@ -57,9 +68,22 @@ def main():
     for c in cluster(df.dropna(), clusters, cmin, cmax+2):
         output.writerow(array(c).tolist())
     file.close(out)
-    reshape(output_csv, clusters)
+    reshape(output_csv, clusters, final_outfile, input_csv)
 
-cleanCSV(input_csv)
-main()
-os.remove("meta.csv")
-os.remove("out.csv")
+def usage():
+    print "\nUsage is cluster.py -i <input.csv> -o <output.csv> -c <choice of k (k-means)>"
+
+if __name__ == "__main__":
+    if len(sys.argv) > 1:
+        main(sys.argv[1:])
+        os.remove("meta.csv")
+        os.remove("out.csv")
+
+'''
+if __name__ == "__main__":
+    if sanitize(sys.argv):
+        main(sys.argv[1:])
+        os.remove("meta.csv")
+        os.remove("out.csv")
+def sanitize(argv):
+'''
